DataManager: Added GCSDataManager and supporting classes
- GCSCraftModel to keep track of basic state information

- Send the correct Guided and Auto mode type for plane
  or copter when issuing GOTO commands

- Put autopilot specifc NS_ENUMs into GCSCraftModes.h

- Created GCSCraftModels.h for easy model importing
