+++ iGCS/Models/Craft/GCSCraftModels.h
@@ -1,0 +1,15 @@
+//
+//  GCSCraftModels.h
+//  iGCS
+//
+//  Created by Andrew Brown on 1/22/15.
+//
+//
+
+#ifndef iGCS_GCSCraftModels_h
+#define iGCS_GCSCraftModels_h
+
+#import "GCSCraftModes.h"
+#import "GCSCraftModel.h"
+
+#endif
